Refactor measurement reporting modules
